aiq: Sync to internal repository
Not merged CLs:
511f51d5 LSC: update split algorithm
4ee39d2a AF: fix focus failure in low light env
25d444ce AF: add config for video mode

Updated CLs:
016e8b34 imx464: update lsc with 20% lsc param
83527089 awb: modify awb_head.h
4fa7d8b7 FakeV4l2Device: fix fakecam can't work caused by commit b06e94d
6f2475d7 CamHwIsp20: release tx/rx buf allocated when prepared
b5dbace7 Update amerge params in imx464 json, based on imx464: update
iqfiles for test 20211206
ebe844f1 ADRC & Amerge: support long frame mode in ixp3x
3fb8cacb Update merge params in json(merge with Amerge: support short
                                     frame)
94dabb8a Amerge: support short frame
b6ea2c48 imx464: update iqfiles for test 20211206
b35fcac9 Amerge: move calculation of expo related reg from hwi to algo
b60f62ff ADRC: move calculation of expo related reg from hwi to algo
0374603d aiq_core: transmit the next frame exposure to the algo group0
69214c43 Adehaze: fix pre_wet bug
f3ea016e Fix some compile errors of uninitialized var
89cec5e7 move rk-camera-module.h to include
e67dc57a group_ae: fix invalid memory access reported by valgrind
b06e94d8 RkStream: fix RkStreamEventPollThread stop stucked
7f9edcef RkAiqCalibDbV2: Fix calib memleak
33f601ae aiq: cac: Fix loss of strength decimal
3775104a bayer3dnr: fix bug for lo_clip_wgt
f52f0551 bayertnr: add json prams, rkaiq & iqfiles should update
together

+++ algos/acac/cac_adaptor.cpp
@@ -36,7 +36,7 @@ namespace RkCam {
 #define BYTES_PER_WORD 4
 #define BITS_PER_WORD  (BITS_PER_BYTE * BYTES_PER_WORD)
 
-#define INTERP_CAC(x0, x1, ratio) static_cast<int>((ratio) * ((x1) - (x0)) + (x0))
+#define INTERP_CAC(x0, x1, ratio) ((ratio) * ((x1) - (x0)) + (x0))
 
 #define DIV_ROUND_UP(n, d) (((n) + (d)-1) / (d))
